% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_parametric.r
\name{define_surv_param}
\alias{define_surv_param}
\title{Define Parametric Distribution}
\usage{
define_surv_param(distribution, ...)
}
\arguments{
\item{distribution}{a parametric survival distribution
(see \code{\link[=define_surv_flexsurv]{define_surv_flexsurv()}} for supported distributions).}

\item{...}{distribution parameters
(see \code{\link[=define_surv_flexsurv]{define_surv_flexsurv()}} for details).}
}
\value{
a \code{surv_parametric} object.
}
\description{
Alias for \code{\link[=define_surv_flexsurv]{define_surv_flexsurv()}}. Define a parametric survival
distribution using flexsurv's native parameterization.
}
\examples{

define_surv_param(distribution = "exp", rate = .5)
define_surv_param(distribution = "gompertz", rate = .5, shape = 1)

}
