% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_hazards.r
\name{max_hazards}
\alias{max_hazards}
\title{Maximum Hazards}
\usage{
max_hazards(dist1, dist2, ..., cycle_length = 1)
}
\arguments{
\item{dist1}{first survival distribution}

\item{dist2}{second survival distribution}

\item{...}{additional survival distributions}

\item{cycle_length}{step size used by \code{surv_quantile} for grid search
(default 1). Has no effect on \code{surv_prob}.}
}
\value{
A \code{surv_max_haz} object.
}
\description{
Combine two or more survival distributions by taking the pointwise
maximum of their discrete hazards at each time step.
}
\examples{

dist1 <- define_surv_param("exp", rate = 0.01)
dist2 <- define_surv_param("exp", rate = 0.02)
max_dist <- max_hazards(dist1, dist2)

}
