% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_quantile.r
\name{surv_quantile}
\alias{surv_quantile}
\title{Evaluate Survival Quantiles}
\usage{
surv_quantile(x, probs, ...)
}
\arguments{
\item{x}{A \code{surv_dist} object}

\item{probs}{A numeric vector of survival probabilities in [0, 1]}

\item{...}{additional arguments passed to methods}
}
\value{
A numeric vector of times
}
\description{
Find the time t at which the survival probability equals a given
value.  That is, find t such that S(t) = p.
}
\examples{
dist1 <- define_surv_param('exp', rate = 0.12)
surv_quantile(dist1, c(0.9, 0.5, 0.1))

}
