% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_random.r
\name{surv_random}
\alias{surv_random}
\title{Generate Random Survival Times}
\usage{
surv_random(x, n, ...)
}
\arguments{
\item{x}{A \code{surv_dist} object}

\item{n}{Number of random samples to generate}

\item{...}{additional arguments passed to \code{surv_quantile}}
}
\value{
A numeric vector of random event times
}
\description{
Sample random event times from any survival distribution using
inverse-transform sampling via \code{surv_quantile}.
}
\examples{
dist1 <- define_surv_param('exp', rate = 0.05)
samples <- surv_random(dist1, 100)

}
